requirements.txt
blinker==1.9.0
click==8.2.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
paho-mqtt==2.1.0
Werkzeug==3.1.3
flask-cors
mkdocs
mkdocs-material
djangorestframework
django-cors-headers
daphne
djoser
google-api-python-client
google-auth-httplib2
google-auth-oauthlib